+++ pytak/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2022 Greg Albrecht <oss@undef.net>
+# Copyright 2023 Greg Albrecht <oss@undef.net>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,12 +16,10 @@
 # Author:: Greg Albrecht W2GMD <oss@undef.net>
 #
 
-"""
-Python Team Awareness Kit (PyTAK) Module.
-~~~~
+"""Python Team Awareness Kit (PyTAK) Module.
 
 :author: Greg Albrecht W2GMD <oss@undef.net>
-:copyright: Copyright 2022 Greg Albrecht
+:copyright: Copyright 2023 Greg Albrecht
 :license: Apache License, Version 2.0
 :source: <https://github.com/ampledata/pytak>
 """
@@ -69,5 +67,5 @@ from . import asyncio_dgram  # NOQA
 
 
 __author__ = "Greg Albrecht W2GMD <oss@undef.net>"
-__copyright__ = "Copyright 2022 Greg Albrecht"
+__copyright__ = "Copyright 2023 Greg Albrecht"
 __license__ = "Apache License, Version 2.0"