+++ pytak/client_functions.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2022 Greg Albrecht <oss@undef.net>
+# Copyright 2023 Greg Albrecht <oss@undef.net>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,17 +21,20 @@
 import argparse
 import asyncio
 import importlib
+import ipaddress
 import logging
 import os
 import platform
+import pprint
 import socket
 import ssl
 import struct
 import sys
+import warnings
 
 from configparser import ConfigParser, SectionProxy
 from urllib.parse import ParseResult, urlparse
-from typing import Any
+from typing import Any, Tuple
 
 import pytak
 
@@ -45,16 +48,16 @@ from pytak.asyncio_dgram import (
 if sys.version_info[:2] >= (3, 7):
     from asyncio import get_running_loop
 else:
+    warnings.warn("Using Python < 3.7, consider upgrading Python.")
     from asyncio import get_event_loop as get_running_loop
 
 __author__ = "Greg Albrecht W2GMD <oss@undef.net>"
-__copyright__ = "Copyright 2022 Greg Albrecht"
+__copyright__ = "Copyright 2023 Greg Albrecht"
 __license__ = "Apache License, Version 2.0"
 
 
-async def create_udp_client(url: ParseResult) -> DatagramClient:
-    """
-    Creates an AsyncIO UDP network client for Unicast, Broadcast & Multicast.
+async def create_udp_client(url: ParseResult) -> Tuple[DatagramClient, DatagramClient]:
+    """Create an AsyncIO UDP network client for Unicast, Broadcast & Multicast.
 
     Parameters
     ----------
@@ -78,9 +81,15 @@ async def create_udp_client(url: ParseResult) -> DatagramClient:
         rsock = reader.socket
         rsock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
         rsock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-    
-    if "239.2.3.1" in host:
-        print("MULTICAST!")
+
+    is_multicast: bool = False
+    try:
+        is_multicast = ipaddress.ip_address(host).is_multicast
+    except ValueError:
+        # It's probably not an ip address...
+        pass
+
+    if is_multicast:
         rsock = reader.socket
         group = socket.inet_aton(host)
         mreq = struct.pack('4sL', group, socket.INADDR_ANY)
@@ -90,8 +99,7 @@ async def create_udp_client(url: ParseResult) -> DatagramClient:
 
 
 def get_tls_config(config: SectionProxy) -> SectionProxy:
-    """
-    Gets the TLS config and ensures required TLS params are set.
+    """Get the TLS config and ensures required TLS params are set.
 
     Parameters
     ----------
@@ -110,8 +118,10 @@ def get_tls_config(config: SectionProxy) -> SectionProxy:
         )
     )
 
-    if None in tls_config_req.values():
-        raise Exception(f"Not all TLS Params specified: {pytak.DEFAULT_TLS_PARAMS_REQ}")
+
+    if not all(tls_config_req.values()):
+        raise Exception(
+            f"Not all required TLS Params specified: {pytak.DEFAULT_TLS_PARAMS_REQ}")
 
     tls_config_opt: dict = dict(
         zip(
@@ -121,15 +131,16 @@ def get_tls_config(config: SectionProxy) -> SectionProxy:
     )
 
     tls_config_req.update(tls_config_opt)
+
     return ConfigParser(dict(filter(lambda x: x[1], tls_config_req.items())))["DEFAULT"]
 
 
-async def protocol_factory(  # pylint: disable=too-many-locals,too-many-branches
+async def protocol_factory(  # NOQA pylint: disable=too-many-locals,too-many-branches,too-many-statements
     config: SectionProxy,
 ) -> Any:
-    """
-    Creates a network connection class instances for the protocol specified by
-    the COT_URL parameter in the config object.
+    """Create a network connection class instance.
+
+    Class is for the protocol specified by the COT_URL parameter in the config object.
 
     Parameters
     ----------
@@ -141,13 +152,13 @@ async def protocol_factory(  # pylint: disable=too-many-locals,too-many-branches
     `Any`
         Return value depends on the network protocol.
     """
-    reader = None
-    writer = None
+    reader: Any = None
+    writer: Any = None
 
-    _cot_url: str = config.get("COT_URL")
+    _cot_url: str = config.get("COT_URL", "")
 
     if "://" not in _cot_url:
-        print("ERROR: Invalid COT_URL: %s", _cot_url)
+        warnings.warn(f"Invalid COT_URL: '{_cot_url}'", SyntaxWarning)
         raise Exception(
             "Please specify COT_URL as a full URL, including '://', for "
             "example: tcp://tak.example.com:1234"
@@ -161,7 +172,7 @@ async def protocol_factory(  # pylint: disable=too-many-locals,too-many-branches
         reader, writer = await asyncio.open_connection(host, port)
     elif scheme in ["tls", "ssl"]:
         host, port = pytak.parse_url(cot_url)
-        tls_config: ConfigParser = get_tls_config(config)
+        tls_config: SectionProxy = get_tls_config(config)
 
         client_cert = tls_config.get("PYTAK_TLS_CLIENT_CERT")
         client_key = tls_config.get("PYTAK_TLS_CLIENT_KEY")
@@ -197,33 +208,34 @@ async def protocol_factory(  # pylint: disable=too-many-locals,too-many-branches
 
         # Default to checking hostnames:
         if dont_check_hostname:
-            print(
-                "WARN: pytak TLS CN/Hostname Check DISABLED by "
-                "PYTAK_TLS_DONT_CHECK_HOSTNAME environment."
-            )
+            warnings.warn(
+                "TLS CN/Hostname Check DISABLED by PYTAK_TLS_DONT_CHECK_HOSTNAME.")
             ssl_ctx.check_hostname = False
 
         # Default to verifying cert:
         if dont_verify:
-            print(
-                "WARN: pytak TLS Certificate Verification DISABLED by "
-                "PYTAK_TLS_DONT_VERIFY environment."
-            )
+            warnings.warn(
+                "TLS Certificate Verification DISABLED by PYTAK_TLS_DONT_VERIFY.")
             ssl_ctx.verify_mode = ssl.CERT_NONE
 
-        reader, writer = await asyncio.open_connection(host, port, ssl=ssl_ctx)
+        try:
+            reader, writer = await asyncio.open_connection(host, port, ssl=ssl_ctx)
+        except ssl.SSLCertVerificationError as exc:
+            raise Exception("Consider setting PYTAK_TLS_DONT_CHECK_HOSTNAME=1 ?") from exc
     elif "udp" in scheme:
         reader, writer = await pytak.create_udp_client(cot_url)
     elif "http" in scheme:
         raise Exception("TeamConnect / Sit(x) Support comming soon.")
         # writer = await pytak.create_tc_client(cot_url)
     elif "log" in scheme:
-        dest: str = cot_url.hostname.lower()
-        if "stderr" in dest:
-            writer = sys.stderr.buffer
-        else:
-            writer = sys.stdout.buffer
-    else:
+        if cot_url.hostname:
+            dest: str = cot_url.hostname.lower()
+            if "stderr" in dest:
+                writer = sys.stderr.buffer
+            else:
+                writer = sys.stdout.buffer
+
+    if not reader and not writer:
         raise Exception(
             "Please specify a protocol in your COT Destination URL. See PyTAK README."
         )
@@ -234,8 +246,7 @@ async def protocol_factory(  # pylint: disable=too-many-locals,too-many-branches
 async def txworker_factory(
     queue: asyncio.Queue, config: SectionProxy
 ) -> pytak.TXWorker:
-    """
-    Creates a PyTAK TXWorker based on URL parameters.
+    """Create a PyTAK TXWorker based on URL parameters.
 
     :param cot_url: URL to COT Destination.
     :param event_queue: asyncio.Queue worker to get events from.
@@ -248,8 +259,7 @@ async def txworker_factory(
 async def rxworker_factory(
     queue: asyncio.Queue, config: SectionProxy
 ) -> pytak.RXWorker:
-    """
-    Creates a PyTAK TXWorker based on URL parameters.
+    """Create a PyTAK TXWorker based on URL parameters.
 
     :param cot_url: URL to COT Destination.
     :param event_queue: asyncio.Queue worker to get events from.
@@ -278,9 +288,40 @@ async def main(app_name: str, config: SectionProxy) -> None:
     await clitool.run()
 
 
+def read_pref_package(pref_package: str) -> dict:
+    """Read a pref package / data package of preferences."""
+    pref_config = {
+        "COT_URL": None,
+        "PYTAK_TLS_CLIENT_CERT": None,
+        "PYTAK_TLS_CLIENT_KEY": None,
+        "PYTAK_TLS_CLIENT_CAFILE": None
+    }
+
+    dp_path: str = pytak.functions.unzip_file(pref_package)
+    pref_file: str = pytak.functions.find_file(dp_path, "*.pref")
+    prefs: dict = pytak.functions.load_preferences(pref_file, dp_path)
+
+    connect_string: str = prefs.get("connect_string", "")
+    assert connect_string
+    pref_config["COT_URL"] = pytak.functions.cs2url(connect_string)
+
+    cert_location: str = prefs.get("certificate_location", "")
+    assert os.path.exists(cert_location)
+
+    client_password: str = prefs.get("client_password", "")
+    assert client_password
+
+    pem_certs: dict = pytak.functions.convert_cert(cert_location, client_password)
+    pref_config["PYTAK_TLS_CLIENT_CERT"] = pem_certs.get("cert_pem_path")
+    pref_config["PYTAK_TLS_CLIENT_KEY"] = pem_certs.get("pk_pem_path")
+    pref_config["PYTAK_TLS_CLIENT_CAFILE"] = pem_certs.get("ca_pem_path")
+
+    assert all(pref_config)
+    return pref_config
+
+
 def cli(app_name: str) -> None:
-    """
-    Abstract implementation of a Command Line Interface (CLI).
+    """Abstract implementation of a Command Line Interface (CLI).
 
     Parameters
     ----------
@@ -298,33 +339,62 @@ def cli(app_name: str) -> None:
         type=str,
         help="Optional configuration file. Default: config.ini",
     )
+    parser.add_argument(
+        "-p",
+        "--PREF_PACKAGE",
+        dest="PREF_PACKAGE",
+        required=False,
+        type=str,
+        help="Optional connection preferences package zip file (aka data package).",
+    )
     namespace = parser.parse_args()
     cli_args = {k: v for k, v in vars(namespace).items() if v is not None}
 
     # Read config:
     env_vars = os.environ
+    
+    # I blame myself for this:
+    env_vars["PS1"] = ""
+    env_vars["PS2"] = ""
+    env_vars["PS3"] = ""
+    env_vars["_PSA"] = ""
+    env_vars["_PSB"] = ""
+    env_vars["_PSC"] = ""
+    env_vars["_PSD"] = ""
+    env_vars["_PSE"] = ""
+    env_vars["C1"] = ""
+    env_vars["C2"] = ""
+    env_vars["CBG"] = ""
+    env_vars["CEND"] = ""
+    env_vars["_PSPre"] = ""
+    env_vars["USERHOST"] = ""
+    env_vars["TMUX_PANE"] = ""
+    env_vars["HISTTIMEFORMAT"] = ""
+
     env_vars["COT_URL"] = env_vars.get("COT_URL", pytak.DEFAULT_COT_URL)
     env_vars["COT_HOST_ID"] = f"{app_name}@{platform.node()}"
     env_vars["COT_STALE"] = getattr(app, "DEFAULT_COT_STALE", pytak.DEFAULT_COT_STALE)
-    config: ConfigParser = ConfigParser(env_vars)
+    _config: ConfigParser = ConfigParser(env_vars)
 
     config_file = cli_args.get("CONFIG_FILE")
     if os.path.exists(config_file):
         logging.info("Reading configuration from %s", config_file)
-        config.read(config_file)
+        _config.read(config_file)
     else:
-        config.add_section(app_name)
+        _config.add_section(app_name)
 
-    config: SectionProxy = config[app_name]
+    config: SectionProxy = _config[app_name]
+
+    pref_package: str = config.get("PREF_PACKAGE", cli_args.get("PREF_PACKAGE"))
+    if pref_package and os.path.exists(pref_package):
+        pref_config = read_pref_package(pref_package)
+        config.update(pref_config)
 
     debug = config.getboolean("DEBUG")
     if debug:
-        import pprint  # pylint: disable=import-outside-toplevel
-
-        # FIXME: This doesn't work with weird bash escape stuff.
         print("Showing Config: %s", config_file)
         print("=" * 10)
-        pprint.pprint(config)
+        pprint.pprint(dict(config))
         print("=" * 10)
 
     if sys.version_info[:2] >= (3, 7):